pyproject.toml
[tool.poetry]
name = "magi-bot"
version = "0.1.0"
description = "A Telegram bot for DeFi and NFT lending information"
authors = ["Sumit Ghosh <sumit@sumit-ghosh.com>"]

[tool.poetry.dependencies]
python = "^3.12,<3.13"
python-telegram-bot = "^20.0"
aiohttp = "^3.8.1"
structlog = "^23.1.0"
langchain = "^0.1.12"
langchain-openai = "^0.0.8"
pydantic-settings = "^2.2.1"
jinja2 = "^3.1.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
